+++ e2e/cypress/support/commands.js
@@ -14,6 +14,7 @@ import {
 import { defaultConfig } from '../testAppConfig';
 import { fireEvent } from '@testing-library/react';
 import RicosDriver from '../../../packages/ricos-driver/web/src/RicosDriver';
+import { ONCHANGE_DEBOUNCE_TIME } from '../../../packages/ricos-editor/web/src/utils/editorUtils';
 import { merge } from 'lodash';
 // Viewport size commands
 const resizeForDesktop = () => cy.viewport('macbook-15');
@@ -107,13 +108,16 @@ Cypress.Commands.add('loadTestAppOnSsr', (fixtureName, compName) => {
 });
 
 Cypress.Commands.add('matchContentSnapshot', () => {
-  if (Cypress.env('MATCH_CONTENT_STATE'))
+  if (Cypress.env('MATCH_CONTENT_STATE')) {
+    cy.wait(ONCHANGE_DEBOUNCE_TIME);
     cy.window()
       .its('__CONTENT_SNAPSHOT__')
       .toMatchSnapshot();
+  }
 });
 
 Cypress.Commands.add('getViewer', () => {
+  cy.wait(ONCHANGE_DEBOUNCE_TIME);
   cy.get('[data-hook="ricos-viewer"]');
 });