+++ e2e/test-env/src/client/RicosTestApp/RicosTestApp.jsx
@@ -40,7 +40,7 @@ class RicosTestApp extends PureComponent {
       ],
     });
 
-    const { contentState, onEditorChange, locale, isMobile, testAppConfig = {} } = this.props;
+    const { contentState, onRicosEditorChange, locale, isMobile, testAppConfig = {} } = this.props;
     const { addPluginMenuConfig, footerToolbarConfig } = testAppConfig.toolbarConfig || {};
     const { skipCssOverride, paletteType } = testAppConfig.theme || {};
     const palette = determinePalette(paletteType);
@@ -54,13 +54,9 @@ class RicosTestApp extends PureComponent {
         theme={palette && createTheme({ palette })}
         cssOverride={!skipCssOverride && theme}
         toolbarSettings={createToolbarSettings(addPluginMenuConfig, footerToolbarConfig)}
+        onChange={onRicosEditorChange}
       >
-        <RichContentEditor
-          config={testAppConfig.pluginsConfig}
-          helpers={{ onVideoSelected }}
-          // using the Ricos onChange causes a delay between the editor and viewer bc of the usage of debounce
-          onChange={onEditorChange}
-        />
+        <RichContentEditor config={testAppConfig.pluginsConfig} helpers={{ onVideoSelected }} />
       </RicosEditor>
     );
   };
@@ -120,7 +116,7 @@ RicosTestApp.propTypes = {
   contentState: PropTypes.object,
   editorState: PropTypes.object,
   localeResource: PropTypes.object,
-  onEditorChange: PropTypes.func,
+  onRicosEditorChange: PropTypes.func,
   seoMode: PropTypes.bool,
   testAppConfig: PropTypes.object,
 };